:recycle: Move Effect in its own file and namespace isEffect
