+++ packages/core/src/index.ts
@@ -1,5 +1,6 @@
 export { cuillere as default } from './cuillere'
 export * from './cuillere'
+export * from './effect'
 export * from './executable-promise'
 export * from './generator'
 export * from './plugins'
