+++ packages/core/src/plugins/plugin.ts
@@ -1,4 +1,5 @@
-import { Operation, Effect } from '../operations'
+import { Effect } from '../effect'
+import { Operation } from '../operations'
 import { GeneratorFunction } from '../generator'
 
 export type HandleFunction<Context = any> = GeneratorFunction<[Operation, Context], any, Effect>
